Add extra test for materilized views and minor refactor.
Creating a materialized view fails if it cannot be executed in a
security-restricted operation so that the creation-time behavior
matches the behavior when the materialized view is refreshed.

Also, slightly refactor ExecCreateTableAs() to enter the
security-restricted operation inside intorel_startup() -- after the
destination object is already created -- to better match the behavior
of REFRESH MATERIALIZED VIEW. This refactor prepares for an upcoming
change fixing the search_path to a safe value during refresh.