CREATE DATABASE: make LOCALE apply to all collation providers.
If LOCALE is specified, it is now the default for ICU_LOCALE in
addition to LC_COLLATE and LC_CTYPE; but any of those can still be set
individually.

If the provider is "builtin", and LOCALE is specified, then it must be
either "C" or "POSIX". If the provider is "builtin" and the locale is
not specified, it is assumed to be "C".

Previously, LOCALE applied only to LC_COLLATE and LC_CTYPE. That could
lead to confusion when the provider is implicit, such as when it is
inherited from the template database, or when ICU was made default at
initdb time in commit 27b62377b4.

Equivalent changes to initdb.

Reverts incomplete fix 5cd1a5af4d.

Discussion: https://postgr.es/m/3391932.1682107209@sss.pgh.pa.us