Simplify collation cache.
Now that the result of pg_newlocale_from_collation() is always
non-NULL, move the collate_is_c and ctype_is_c flags into pg_locale_t,
and always use that. This commit eliminates the multi-stage
initialization of the cache and the extra code in lc_collate_is_c()
and lc_ctype_is_c().