downcase_identifier(): use method table from locale provider.
Refactor to allow each provider to supply its own implementation of
identifier casefolding.

When the database default locale is libc, use tolower_l() instead of
tolower() to remove its dependency on the global LC_CTYPE setting.

When the database default locale is ICU, and the encoding is
single-byte, tolower() is still required to preserve historical
behavior. We should consider fixing this separately.

One minor behavior change is that, before the database default locale
is initialized, it uses ASCII semantics to fold the
identifiers. Previously, it would use the postmaster's LC_CTYPE
setting from the environment. While that could have some effect during
GUC processing, for example, it would have been fragile to rely on the
environment setting anyway. (Also, it only matters when the encoding
is single-byte.)